+++ python_modules/dagster-webserver/dagster_webserver_tests/test_smoke.py
@@ -1,5 +1,5 @@
 import pytest
-from dagster._cli.workspace.cli_target import WorkspaceOpts
+from dagster._cli.workspace.cli_target import WorkspaceOpts, workspace_opts_to_load_target
 from dagster._core.test_utils import instance_for_test
 from dagster._core.workspace.context import WorkspaceProcessContext
 from dagster_webserver import app
@@ -34,9 +34,9 @@ def test_smoke_app(gen_instance):
             instance,
             version="",
             read_only=False,
-            workspace_load_target=WorkspaceOpts(
-                module_name=("dagster_webserver_tests.toy.bar_repo",)
-            ).to_load_target(),
+            workspace_load_target=workspace_opts_to_load_target(
+                WorkspaceOpts(module_name=("dagster_webserver_tests.toy.bar_repo",))
+            ),
         ) as workspace_process_context:
             asgi_app = app.create_app_from_workspace_process_context(workspace_process_context)
             client = TestClient(asgi_app)
