+++ python_modules/dagster/dagster/_cli/code_server.py
@@ -7,9 +7,10 @@ from typing import Optional
 
 import click
 import dagster_shared.seven as seven
+from dagster_shared.cli import python_pointer_options
 
 from dagster._cli.utils import assert_no_remaining_opts
-from dagster._cli.workspace.cli_target import PythonPointerOpts, python_pointer_options
+from dagster._cli.workspace.cli_target import PythonPointerOpts
 from dagster._core.instance import InstanceRef
 from dagster._core.types.loadable_target_origin import LoadableTargetOrigin
 from dagster._core.utils import FuturesAwareThreadPoolExecutor
